requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
django-js-asset==2.0.0
django-model-utils==4.2.0
django-mptt==0.13.4
djangorestframework==3.13.1
docopt==0.6.2
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.0
packaging==21.3
Pillow==9.0.1
pipreqs==0.4.11
psycopg2==2.9.3
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
w3lib==1.22.0
yarg==0.1.9
